5.0.0: Allow EmailTransaction to be configured to set SMTP headers
For example, many external email providers require authentication
headers to be sent with emails that are sent.

These headers are logically part of the connection to the email
provider, and not part of the emails themselves, that's why they're set
in the EmailTransaction and not by clients on the email objects
themselves.